[Robot Control] refactor the control framework for better understanding
