+++ aerial_robot_control/include/aerial_robot_control/control/fully_actuated_controller.h
@@ -35,7 +35,7 @@
 #pragma once
 
 #include <aerial_robot_msgs/WrenchAllocationMatrix.h>
-#include <aerial_robot_control/control/pose_linear_controller.h>
+#include <aerial_robot_control/control/base/pose_linear_controller.h>
 #include <spinal/FourAxisCommand.h>
 #include <spinal/RollPitchYawTerms.h>
 #include <spinal/TorqueAllocationMatrixInv.h>
