+++ aerial_robot_control/include/aerial_robot_control/control/under_actuated_controller.h
@@ -34,7 +34,7 @@
  *********************************************************************/
 #pragma once
 
-#include <aerial_robot_control/control/pose_linear_controller.h>
+#include <aerial_robot_control/control/base/pose_linear_controller.h>
 #include <spinal/FourAxisCommand.h>
 #include <spinal/RollPitchYawTerms.h>
 #include <spinal/TorqueAllocationMatrixInv.h>
