+++ aerial_robot_control/src/control/base/pose_linear_controller.cpp
@@ -34,7 +34,7 @@
  *********************************************************************/
 
 
-#include <aerial_robot_control/control/pose_linear_controller.h>
+#include <aerial_robot_control/control/base/pose_linear_controller.h>
 
 namespace aerial_robot_control
 {
